# Threaded polling of connections