Added special variable configuration so tests will probably fail because it's not in the wiki.