@Safihre 2023-01-31
@mnightingale do you think we even need multiple decoder threads this way?