@mnightingale 2023-01-31
Well I was confused for way too long until I noticed it defaults to 1 thread 🤣 - would suggest a minimum based on `os.cpu_count()`  - `min(4, int(os.cpu_count() / 2))` or something.
I think this is a neater approach, it does have the drawbacks that I outlined in the other PR just that most of the time you'll only get very few sockets ready to read at once so its unlikely to utilise many threads.

> @mnightingale do you think we even need multiple decoder threads this way?

I'm unconvinced multiple decoders or a separate thread for them is necessary regardless, from all I've played with encoding and decoding it's extremely fast, much faster than its possible to download. But it would need investigation, it's possible there is a lot of overhead calling the function and releasing the GIL - side thought I wonder if there is a way of creating a buckets of articles to pass to the decoder rather than one at a time to avoid any Python=>C or releasing and reacquiring the GIL overhead, which may actually be quicker than doing one at a time while downloading.

That is as long as SIMD is being used which even on slow hardware can do hundreds of MB/s, I presume the multiple workers was from before SIMD decoding was added. But then I've never tried it on RPi or anything so I could be wrong.

But to answer your question I think this way of using a ThreadPool is likely to often result in only reading from 1 or 2 sockets at a time so I think it would be very similar performance to the current queue/worker approach, it would probably be a good idea to analyse the frequency of sockets against a real server, if it's 1 a lot of the time the queue/worker approach may be better.

The decode would need to be handled within the pool, i.e. not in the `__handle_recv_result` so you'd probably want to request the next article before you start decoding, so the server can start sending it while you decode the previous one.

@puzzledsab looking at `__handle_recv_result` I noticed the `BPSMeter` sleeping, I think it should come after all the `imap_unordered` results are all in not for each one.

Additionally and even if this PR isn't merged for some reason maybe I've misunderstood it's purpose but if you're hitting the bandwidth limit it's going to wait until it's below before queuing finished articles for decoding, I think it'd be better to queue it for the decoder to work on while we sleep, perhaps the intention was to sleep before it requests the next article? - in which case I think the "if done ... decode" part should be before the possible sleeps.