@puzzledsab 2023-02-01
1 thread is default because then it will work like before, not using Threadpool at all. I was hoping that would make it easier to accept and I also think it may work better with the speed limiter, although maybe not if the sockets have huge receive buffers.

The threading is primarily intended for Rpi 4, possibly other cheap multi core CPUs and anyone with extremely fast connections and a good number of cores. On the Rpi 4 it's possible to get a decent speedup by using stunnel (SSL proxy) and unencrypted transfer in SABnzbd. It may be partly because then it can request more than 16 KB but I think it's also because it will spread the decryption load on more cores. Using multiple threads for reading in combination with the new receive code will hopefully do the same.

I don't think overhead related to GIL switching is a big part for the decoder threads. The current method for sleeping doesn't work all that well. Usually it will trigger at most every other round, probably less, and receive 16KB or less from every socket like it did in the old version. `unlocked_ssl_recv_into()` may be called 25-50 times for decoder call so that's where the overhead would be. I did see someone report improved performance by increasing the decoder threads to 3 or 4 with the current optimized decoder.

I hope it will at least be included in the beta so we can get some real world tests to see if it works.