@mnightingale 2023-02-04
I finally managed to get a RPi 4, testing with my fake server over the local network initially.
The ThreadPool does appear to make a noticeable difference, from 30MB/s to 50MB/s with 2 threads but only 60 MB/s with 4.
I only test with 8 connections though for consistency so it'll be rare that many threads are used.

I also tried against a real server and went 33-42ish (jumps around a bit), but that's more or less my line speed, hence I prefer testing with a fake server.

It's really just convincing me that a rework with a thread per connection may be worth the effort.

@puzzledsab I believe you use a RPi 3? what kind of difference have you noticed?