@puzzledsab 2023-02-04
Cool! The Rpi 3 only supports 100 Mbit/s so it could always fill the maximum bandwidth of about 11 MB/s. The CPU usage according to top with the current develop branch has been reduced by just over 20%, though.

I merged the develop branch to fix a conflict so now it will also have the new sleep method. That might help slightly, particularly if you increase the `downloader_sleep_time` to maybe 50 (= 5 ms) or more. On my Rpi there seems to be some extra overhead when sleeping the default of 2 ms or less.

Did you try to increase the decoder threads? Did the cache fill so that it started swapping parts to disk? Did you actually write the data to disk or disable that part? Have you tried a different number of connections? What is your CPU usage according to top? If it's close to 400 with the new sleep then I don't think there's much more to do with the current design.

This is quite exciting. It would be nice to get even more than 60 MB/s but a doubling is still pretty cool. Hopefully it works in the real world as well. It's all coming together with the new SSL receive method just like we hoped :)