@Safihre 2023-02-05
Great work, will look at this in more detail how we can structure it best. 

> It's really just convincing me that a rework with a thread per connection may be worth the effort.

You really mean a (Python) thread per connection? Not the previous approach of keeping it around a number of threads similar to the number of cores? Again those examples of users actually putting in the 50-100 connections that providers nowadays offer.