@puzzledsab 2023-02-05
Come to think of it, maybe it's better to set `downloader_sleep_time` to 0 in your case if the downloader thread is the bottleneck and it's already filling the receive buffers every time.