@mnightingale 2023-02-05
@puzzledsab I forgot about the 3's limitations, I have my 4 booting and downloading to an SSD via USB, seems to manage 200-300 MB/s in benchmarks (SABs test is nearer 100 MB but I don't think it's at all accurate for fast drives), cipher set to AES128.
It barely hits the cache ~50 articles at most but perhaps eventually I can look at HDDs performance, just focussing on CPU for now.

This is with 16 connections without any other changes from the defaults except thread count:
Using `top -b -c -p $(pgrep -d',' -f SABnzbd) > single.log`

**Single thread:** 
Barely using multiple cores ~130%
Downloaded in 1 min 11 seconds at an average of 28.5 MB/s
https://gist.github.com/mnightingale/870a3a07fd1c97f1a7ed1a0d77ddc6d3

**Four threads:**
Hitting all cores ~330%
Downloaded in 29 seconds at an average of 68.7 MB/s
https://gist.github.com/mnightingale/964195210d79d542c0a58c154913d38c

**Master/3.7.1:**
Just for comparison to the current release branch
Similar to develop, 130% with one core at 100%
Downloaded in 1 min 18 seconds at an average of 26.2 MB/s
https://gist.github.com/mnightingale/b6302f3433b0e084526cde73f972a157

It's quite an improvement but I want to get it to 1 Gbps 😄

---

> You really mean a (Python) thread per connection? Not the previous approach of keeping it around a number of threads similar to the number of cores? Again those examples of users actually putting in the 50-100 connections that providers nowadays offer.

Honestly I'm not sure; if I put together a POC I'll probably do a thread per connection, it's by far the simplest to implement, the threads would need to be very tight, [selectors](https://docs.python.org/3/library/selectors.html) would be used with no timeout so threads return as soon as possible, maybe some waiting events to avoid listing on known idle threads at all but the select calls are system/kernel levels calls so extremely fast anyway.

From what I've read Python is good at scheduling threads, forcing a GIL holding thread to release every 5ms and the actual time consuming part of download/decode releases the GIL anyway and requests it again when needed.

Certainly I think pools of connections have potential, the complicated bit would be balancing those connections - probably round robin each priority and handling changes, but the actual logic of the thread/worker would be very similar whether it's dealing with 1 connection or many, hence implementing a thread per connection first then investigating if it's causing issues and possibly create pools seems like the way to go.

---

I highly recommend checking out the book [High Performance Python, 2nd Edition](https://www.oreilly.com/library/view/high-performance-python/9781492055013/) I only started reading it over the weekend and I'm only a few chapters in but so far it's excellent and there are a lot of topics that are very relevant to SAB.