@mnightingale 2023-02-08
On a RPi 4 changing the cipher to ECDHE-RSA-CHACHA20-POLY1305 helped a lot.

This is with 16 connections:

Master: 79.0 MB/s
(no image)

Develop: 102.7 MB/s
![image](https://user-images.githubusercontent.com/9887246/217272394-24dbc7e8-1a40-4145-891b-c8ecdd4e2972.png)

Threadpool: 107.7 MB/s (2 threads)
![image](https://user-images.githubusercontent.com/9887246/217272904-dc20719f-5670-4cc8-bde0-2c637180b664.png)

CPU usage goes from ~230 to ~260 so still plenty to spare, the bandwidth graph is a bit more stable, and it's pretty much hitting 1 Gbps which is quite impressive for such a device.[^1]

Therefore I think for at least the medium term this approach is a good improvement.

For the longer term more investigation into possible approaches in Python are probably required; threads, threadpools or something else however I think for most users it could just be a lot of work for little gain.[^2]
Because I presume you want to get to a beta release stage I think it's best to use this PR and maybe rework internals for a future release, I wouldn't want a beta/release to be blocked by such a change.

[^1]: It did make me wonder if we could detect if we are running on a Raspberry Pi and override the default cipher, it's pretty complicated for users to go into `openssl speed` and make any sense of it.

[^2]: My current line of thinking is:
**threads** - probably fastest but would need mechanisms for back-pressure, probably similar to issues that are dealt with currently.
**threadpools** - likely similar performance to threads, more complicated to manage (spread out the sockets), less threads for Python to manage/schedule.
**others** - I don't know what I don't know, I tried asyncio but it seems to be more for concurrency than parallel, it seemed really efficient and you can write a really clean protocol for interacting with nntp. But its was just managing one thread so it's slower than threading even more so when I added decoding, but this could just be because I don't know how to use it well.
**tl;dr** - I think threads are probably where I'd start.